moving state related things out of composetools/
